refactor(auth): Phase 2 move User, AuthToken to Domain/
- Relocate User + AuthToken records to Domain/ (namespace
  Auth.Service.Domain); ServiceClient stays in Models/ per phase scope.
- Update consumer usings across Auth.Service + Auth.Tests; refresh EF
  migration snapshot entity-name strings to the new namespace.
