+++ auth-microservice/Auth.Service/Domain/AuthToken.cs
@@ -1,3 +1,3 @@
-namespace Auth.Service.Models;
+namespace Auth.Service.Domain;
 
 public record AuthToken(string Token, int ExpiresIn);
