+++ auth-microservice/Auth.Service/Domain/User.cs
@@ -1,4 +1,4 @@
-namespace Auth.Service.Models;
+namespace Auth.Service.Domain;
 
 public class User
 {
