+++ auth-microservice/Auth.Service/Services/ITokenService.cs
@@ -1,4 +1,4 @@
-using Auth.Service.Models;
+using Auth.Service.Domain;
 
 namespace Auth.Service.Services;
