+++ auth-microservice/Auth.Service/Services/JwtTokenService.cs
@@ -1,8 +1,8 @@
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
+using Auth.Service.Domain;
 using Auth.Service.Infrastructure.Data;
-using Auth.Service.Models;
 using Auth.Service.Services.Signing;
 using ECommerce.Shared.Authentication;
 using Microsoft.IdentityModel.Tokens;
