+++ auth-microservice/Auth.Service/Services/ServiceTokenService.cs
@@ -2,7 +2,7 @@ using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Security.Cryptography;
 using System.Text;
-using Auth.Service.Models;
+using Auth.Service.Domain;
 using Auth.Service.Services.Signing;
 using ECommerce.Shared.Authentication;
 using Microsoft.IdentityModel.Tokens;
