+++ auth-microservice/Auth.Tests/AuthApiEndpointsTests.cs
@@ -1,7 +1,7 @@
 using System.Diagnostics.Metrics;
 using Auth.Service.ApiModels;
+using Auth.Service.Domain;
 using Auth.Service.Endpoints;
-using Auth.Service.Models;
 using Auth.Service.Services;
 using ECommerce.Shared.Observability.Metrics;
 using Microsoft.AspNetCore.Http.HttpResults;
